pyproject.toml
[tool.poetry]
name = "fastapi-hotels"
version = "0.1.0"
description = ""
authors = ["Nina Nikitenko"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic-settings = ">=2.11.0,<3.0.0"
sqlalchemy = ">=2.0.44,<3.0.0"
alembic = ">=1.17.0,<2.0.0"
asyncpg = ">=0.30.0,<0.31.0"
black = ">=25.9.0,<26.0.0"
pwdlib = { version = ">=0.3.0,<0.4.0", extras = ["argon2"] }
pyjwt = { version = ">=2.10.1,<3.0.0", extras = ["crypto"] }
pydantic = { version = ">=2.12.5,<3.0.0", extras = ["email"] }
redis = ">=7.3.0,<8.0.0"
fastapi-cache2 = ">=0.2.2,<0.3.0"
celery = ">=5.6.2,<6.0.0"
python-multipart = "^0.0.22"
pillow = "^12.1.1"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-env = "^1.6.0"
pytest-asyncio = "^1.3.0"
ruff = "^0.15.9"

[tool.pytest_env]
env_files = [".env.test"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
